Проверка длинны сессии пользователя #78
- добавлена функция для определении длинны сессии пользователей
checkLength
- перенесли функцию определения голосового сообщения checkVoice
 из мейна в файл checks.js
