+++ src/mw.js
@@ -13,13 +13,6 @@ export const startMW = (bot) => {
 
   bot.use(async (ctx, next) => {
     try {
-      // письмо пользователю МАРТ
-      // if (ctx?.message?.from?.id === 6083583477) {
-      //   await ctx.reply('Здравствуйте, Март, это программист данного бота - Карпов Максим! ');
-      //   await ctx.reply('Эту запись добавил специально для Вас!');
-      //   await ctx.reply('Дело в том, что этот бот изначально создавался для меня и моих знакомых, а сейчас я начал замечать, что он используется большим количеством людей, что совсем не планировалось. Поэтому в ближайшее время я планирую включить ограничение доступа. Если вы меня знаете, или просто хотели бы продолжить использовать бота, напишите мне в телегу или вацап, мой номер 8-906-598-71-86. Вы вроде тоже занимаетесь программированием, поэтому просто так не хочу вас удалять ');
-      // }
-
       console.log(`Пользователь в разрешенном списке? - ${accessIsAllowed(ctx?.message?.from?.id)}`);
       // accessIsAllowed(ctx?.message?.from?.id);
